refactor(app): Batch 4 route-local _components/_lib + Discord drilldown fixes
- Move draft/[id] pieces into _components; draft-history clients and drilldown
  into _components/_lib; update imports and path-sensitive tests.
- Fix H2H empty copy to match All vs Captain tab.
- Keep conflict UI on failed Discord claim submit (clear conflict ID only on success).
