+++ src/app/draft/[id]/_components/DraftClient.tsx
@@ -3,7 +3,7 @@
 import { useQuery } from "convex/react";
 import { api } from "../../../../../convex/_generated/api";
 import { DraftData, CurrentPlayer } from "../../types";
-import DraftBoard from "../components/DraftBoard";
+import DraftBoard from "./DraftBoard";
 
 export default function DraftClient({
   shortId,
