+++ src/app/draft/[id]/_components/PlayerAvatar.tsx
